backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
nltk==3.8.1
scikit-learn==1.2.2
numpy==1.24.2
openai==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.1
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
Werkzeug==2.0.1
SQLAlchemy==1.4.23
